refactor: moved options into model.d.ts
