+++ types/data/index.d.ts
@@ -1,2 +1,1 @@
 export * from "./model";
-export * from "./options";
