+++ types/data/model.d.ts
@@ -1,4 +1,18 @@
-import { Options } from "./options";
+export interface Options {
+    readonly fields?: string[];
+    readonly eager?: boolean;
+    readonly eagerL?: boolean;
+    readonly map?: boolean;
+    readonly rules?: boolean;
+    readonly meta?: boolean;
+    readonly build?: boolean;
+    readonly fill?: boolean;
+    readonly resolveA?: boolean;
+    readonly skip?: number;
+    readonly limit?: number;
+    readonly sort?: unknown[][];
+    readonly raiseE?: boolean;
+}
 
 export declare class Model {
     static niw<T = Model>(this: { new (): T }): T;
