--- types/data/options.d.ts
@@ -1,15 +1,0 @@
-export interface Options {
-    readonly fields?: string[];
-    readonly eager?: boolean;
-    readonly eagerL?: boolean;
-    readonly map?: boolean;
-    readonly rules?: boolean;
-    readonly meta?: boolean;
-    readonly build?: boolean;
-    readonly fill?: boolean;
-    readonly resolveA?: boolean;
-    readonly skip?: number;
-    readonly limit?: number;
-    readonly sort?: unknown[][];
-    readonly raiseE?: boolean;
-}
